requirements.txt
fiona==1.10.1
geopandas==1.1.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
pyproj==3.7.2
PyYAML==6.0.3
scipy==1.16.3
Shapely==2.1.2
streamlit==1.51.0
streamlit-image-coordinates==0.4.0